Refactor and enhance code structure across multiple files
- Removed unnecessary blank lines and improved code formatting for better readability in CreateProductData, UpdateProductData, CreateUserData, LoginUserData, and various model files.
- Updated enum declaration style in OrderStatus.
- Adjusted the order of use statements in Category, Order, OrderItem, Product, Stock, and User models for consistency.
- Implemented authorization checks in CategoryService for create, update, and delete operations.
- Introduced Abilities class to define user permissions and integrated it into AuthService for token generation.
- Created CategoryPolicy to manage authorization logic for category actions.
- Updated routes to enforce authentication middleware for category management.
- Added BaseService to handle authorization checks across services
- Enhanced error handling in various controllers and services to provide clearer responses.
- Updated migration files to use consistent formatting and removed unnecessary use statements.
- Added seeder files for categories, orders, products, and stocks without model events.

+++ app/Services/CategoryService.php
@@ -5,13 +5,17 @@ namespace App\Services;
 use App\DTOs\Categories\CreateCategoryData;
 use App\DTOs\Categories\UpdateCategoryData;
 use App\Entities\Models\Category;
+use App\Policies\V1\CategoryPolicy;
 use App\Repositories\Contracts\CategoryRepositoryInterface;
-use Illuminate\Pagination\LengthAwarePaginator;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
+use Illuminate\Pagination\LengthAwarePaginator;
 
-class CategoryService
+class CategoryService extends BaseService
 {
     protected $categoryRepository;
+
+    protected $policyClass = CategoryPolicy::class;
 
     public function __construct(CategoryRepositoryInterface $categoryRepository)
     {
@@ -20,21 +24,25 @@ class CategoryService
 
     public function createCategory(CreateCategoryData $categoryData): Category
     {
+        if (! $this->isAble('store', Category::class)) {
+            throw new AuthorizationException('You do not have permission to create a category.');
+        }
+
         return $this->categoryRepository->create($categoryData);
     }
 
     public function getAllCategories(): LengthAwarePaginator
     {
         return $this->categoryRepository
-                    ->applyQuery()
-                    ->paginate();
+            ->applyQuery()
+            ->paginate();
     }
 
     public function getCategoryById(int $id): Category
     {
         $category = $this->categoryRepository->findById($id);
 
-        if (!$category) {
+        if (! $category) {
             throw new ModelNotFoundException("Category with ID {$id} not found");
         }
 
@@ -43,11 +51,19 @@ class CategoryService
 
     public function updateCategory(UpdateCategoryData $categoryData): Category
     {
+        if (! $this->isAble('update', Category::class)) {
+            throw new AuthorizationException('You do not have permission to update a category.');
+        }
+
         return $this->categoryRepository->update($categoryData);
     }
 
     public function deleteCategory(Category $category): bool
     {
+        if (! $this->isAble('delete', Category::class)) {
+            throw new AuthorizationException('You do not have permission to delete a category.');
+        }
+
         return $this->categoryRepository->delete($category);
     }
 }
